requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.15.4
motor==3.5.0
pydantic==1.10.12
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
fastapi-cors==0.0.6
pyjwt==2.10.1
requests==2.31.0
httpx==0.25.2
numpy>=1.26.0
google-generativeai==0.4.0
